first commit
Given an array of positive integers of size n and a positive integer t "threshold". Divide the array in non-empty subarrays of length atmost equal to t. Means none of the subarrays can be of length greater than t. You have to find the minimum score of the array. Where score is the sum of the maximum integer from each subarray.